refactor(dpp): move shielded sighash preimage builders into their own file (#3816)
Per review feedback, split the consensus-critical platform-sighash preimage code
out of the catch-all `shielded/mod.rs` into a dedicated `shielded/sighash.rs`:
SIGHASH_DOMAIN, compute_platform_sighash, the three `*_extra_sighash_data`
version dispatchers + their `_v0` impls, and their byte-layout tests. The public
paths are unchanged (re-exported from `shielded/mod.rs`), so no caller changes.
`mod.rs` keeps the storage-byte fee constants and the OrchardBundleParams /
SerializedAction wire structs. Pure code move — no logic change; the 7 sighash
tests pass unchanged as `shielded::sighash::tests::*` (166 shielded tests green).
